+++ packages/solana/src/connectors/walletAdapters.ts
@@ -5,18 +5,27 @@ import {
   TrustWalletAdapter
 } from './walletAdapters/index.js'
 
-import type { BaseWalletAdapter } from '@solana/wallet-adapter-base'
+import { WalletAdapterNetwork, type BaseWalletAdapter } from '@solana/wallet-adapter-base'
+import { solana, solanaDevnet, solanaTestnet } from '../utils/chains.js'
 import type { Connector } from '@web3modal/scaffold'
 
 export type AdapterKey = 'phantom' | 'solflare' | 'trustWallet' | 'backpack'
 export const supportedWallets: AdapterKey[] = ['phantom', 'solflare', 'trustWallet', 'backpack']
 
-export function createWalletAdapters() {
+const chainMap = {
+  [solana.chainId]: WalletAdapterNetwork.Mainnet,
+  [solanaDevnet.chainId]: WalletAdapterNetwork.Devnet,
+  [solanaTestnet.chainId]: WalletAdapterNetwork.Testnet
+}
+
+export function createWalletAdapters(chainId?: string) {
   return {
     phantom: new PhantomWalletAdapter(),
     trustWallet: new TrustWalletAdapter(),
     backpack: new BackpackWalletAdapter(),
-    solflare: new SolflareWalletAdapter()
+    solflare: new SolflareWalletAdapter({
+      network: chainMap[chainId || solana.chainId]
+    })
   }
 }