+++ packages/ui/src/components/wui-icon/index.ts
@@ -72,6 +72,7 @@ import { walletPlaceholderSvg } from '../../assets/svg/wallet-placeholder.js'
 import { walletSvg } from '../../assets/svg/wallet.js'
 import { walletConnectSvg } from '../../assets/svg/walletconnect.js'
 import { warningCircleSvg } from '../../assets/svg/warning-circle.js'
+import { idSvg } from '../../assets/svg/id.js'
 import { xSvg } from '../../assets/svg/x.js'
 
 const svgOptions: Record<IconType, TemplateResult<2>> = {
@@ -112,6 +113,7 @@ const svgOptions: Record<IconType, TemplateResult<2>> = {
   github: githubSvg,
   google: googleSvg,
   helpCircle: helpCircleSvg,
+  id: idSvg,
   infoCircle: infoCircleSvg,
   mail: mailSvg,
   mobile: mobileSvg,