+++ examples/react-wagmi/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @examples/react-wagmi
 
+## 4.2.3
+
+### Patch Changes
+
+- feat: - feat: restricted ens names. fix: iat set automatically if not present in messageParams. Adds siwe config handlers
+
+- Updated dependencies []:
+  - @web3modal/wagmi@4.2.3
+
+## 4.2.3-alpha.0
+
+### Patch Changes
+
+- feat: add support for coinbase smart accounts
+
+- Updated dependencies []:
+  - @web3modal/wagmi@4.2.3-alpha.0
+
+## 4.2.2
+
+### Patch Changes
+
+- feat: social login refactor. wagmi sendCalls support. refactor theme variables
+
+- Updated dependencies []:
+  - @web3modal/wagmi@4.2.2
+
 ## 4.2.1
 
 ### Patch Changes